Moved urls to domain
