+++ src/xmipp3_installer/application/logger/errors.py
@@ -1,5 +1,7 @@
 """### Submodule containing all constants needed for handling errors during Xmipp's installation."""
 
+from xmipp3_installer.domain.urls import CMAKE_INSTALL_DOCS_URL
+
 # Error codes
 INTERRUPTED_ERROR = -1
 OK = 0
@@ -15,13 +17,12 @@ IO_ERROR = 7
 LOG_FILE = 'compilation.log'
 
 # Error messages
-__CMAKE_INSTALL_DOCS_URL = 'https://i2pc.github.io/docs/Installation/InstallationNotes/index.html#cmake'
 __CHECK_LOG_MESSAGE = f'Check the inside file \'{LOG_FILE}\'.'
 ERROR_CODES = {
 	INTERRUPTED_ERROR: ['Process was interrupted by the user.', ''],
 	UNKOW_ERROR: ['', ''],
 	SOURCE_CLONE_ERROR: ['Error cloning xmipp repository with git.', 'Please review the internet connection and the git package.'],
-	CMAKE_ERROR: ['There was an error with CMake.', f'Please install it by following the instructions at {__CMAKE_INSTALL_DOCS_URL}'],
+	CMAKE_ERROR: ['There was an error with CMake.', f'Please install it by following the instructions at {CMAKE_INSTALL_DOCS_URL}'],
 	CMAKE_CONFIGURE_ERROR: ['Error configuring with CMake.', __CHECK_LOG_MESSAGE],
 	CMAKE_COMPILE_ERROR: ['Error compiling with CMake.', __CHECK_LOG_MESSAGE],
 	CMAKE_INSTALL_ERROR: ['Error installing with CMake.', __CHECK_LOG_MESSAGE],
