--- src/xmipp3_installer/application/logger/urls.py
@@ -1,3 +1,0 @@
-"""### URLs """
-
-DOCUMENTATION_URL = 'https://i2pc.github.io/docs/'
