+++ src/xmipp3_installer/domain/urls.py
@@ -1,0 +1,4 @@
+"""### URLs """
+
+DOCUMENTATION_URL = 'https://i2pc.github.io/docs/'
+CMAKE_INSTALL_DOCS_URL = 'https://i2pc.github.io/docs/Installation/InstallationNotes/index.html#cmake'
